% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_merges.R
\name{apply_merges}
\alias{apply_merges}
\title{Apply merges based on membership and K* table}
\usage{
apply_merges(
  current_data,
  lineage_dt,
  kstar_all,
  final_membership_list,
  verbose = TRUE
)
}
\description{
Apply merges based on membership and K* table
}
\keyword{internal}
