% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{icdmerge-package}
\alias{icdmerge}
\alias{icdmerge-package}
\title{icdmerge package}
\description{
Tools for merging fine-grained ICD diagnosis nodes into clinically interpretable merged nodes using block-wise consensus k-means clustering, PAC-based K selection, and an optional homogeneity gate.
}
\author{
\strong{Maintainer}: Shui-Kit Cheuk \email{2311210077@bjmu.edu.cn}

}
\keyword{internal}
