% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_inputs.R
\name{prepare_inputs}
\alias{prepare_inputs}
\title{Prepare inputs: prevalence filter, mapping, sparse matrix}
\usage{
prepare_inputs(
  raw_data,
  codebook,
  prevalence_cutoff,
  block_col = "block",
  verbose = TRUE
)
}
\description{
Prepare inputs: prevalence filter, mapping, sparse matrix
}
\keyword{internal}
